Terminé de imitar el nuevo modelo de datos en el frontend
