+++ lib/api/hello_request.dart
@@ -1,4 +1,3 @@
-import 'package:http/http.dart';
 import 'package:ut_report_generator/api/send_request.dart';
 
 Future<HelloRequest_Response> helloRequest() {
