+++ lib/api/report/export_report.dart
@@ -1,10 +1,10 @@
-import 'package:ut_report_generator/api/render_report.dart';
+import 'package:ut_report_generator/api/report/render_report.dart';
 import 'package:ut_report_generator/api/send_request.dart';
 
-Future<ExportReport_Response> exportReport({required String reportDirectory}) {
+Future<ExportReport_Response> exportReport({required String identifier}) {
   return sendRequest(
-    route: "export_report",
-    body: {"report_directory": reportDirectory},
+    route: "report/export",
+    body: {"report": identifier},
     callback: ExportReport_Response.fromJson,
   );
 }
