+++ lib/api/report/import_report.dart
@@ -1,9 +1,9 @@
 import 'package:ut_report_generator/api/send_request.dart';
-import 'package:ut_report_generator/models/report_class.dart';
+import 'package:ut_report_generator/models/report.dart';
 
 Future<ReportClass> importReport({required String reportFile}) async {
   return sendRequest(
-    route: "import_report",
+    route: "report/import",
     callback: ReportClass.fromJson,
     body: {"report_file": reportFile},
   );
