+++ lib/models/image_slide/image_slide_kind.dart
@@ -1,0 +1,1 @@
+enum ImageSlideKind { coverPage, finalPage }
