+++ lib/models/pivot_table/aggregate_function_type.dart
@@ -1,0 +1,1 @@
+enum AggregateFunctionType { count, average, min, max }
