+++ lib/models/pivot_table/filter_function_type.dart
@@ -1,0 +1,1 @@
+enum FilterFunctionType { failedStudents, approvedStudents, allStudents }
