+++ lib/models/slide/self.dart
@@ -1,0 +1,3 @@
+abstract class Slide {
+  String get identifier;
+}
