+++ lib/models/slide_category.dart
@@ -1,0 +1,1 @@
+enum SlideCategory { pivotTable, image }
