+++ lib/pages/home/report-editor/image_slide_section/widget.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:ut_report_generator/models/image_slide/self.dart';
+
+class ImageSlideSection extends StatefulWidget {
+  final ImageSlide initialSlide;
+  const ImageSlideSection({super.key, required this.initialSlide});
+
+  @override
+  State<ImageSlideSection> createState() => _ImageSlideSectionState();
+}
+
+class _ImageSlideSectionState extends State<ImageSlideSection> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
