+++ lib/pages/home/report-editor/pivot_table_section/pick_file_button.dart
@@ -2,8 +2,8 @@ import 'dart:io';
 
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
-import 'package:ut_report_generator/api/change_slide_data.dart';
-import 'package:ut_report_generator/api/start_report.dart';
+import 'package:ut_report_generator/api/image_slide/change_slide_data.dart';
+import 'package:ut_report_generator/api/report/start_report_with_pivot_table.dart';
 
 class PickFileButton extends StatefulWidget {
   String message;
