+++ python-app/routes/report/__init__.py
@@ -4,7 +4,7 @@ from routes.report.get_report import get_report
 from routes.report.import_report import import_report
 from routes.report.render_report import render_report
 from routes.report.start_report_with_pivot_table import start_report_with_pivot_table
-from routes.report.start_report_with_slide import start_report_with_slide
+from routes.report.start_report_with_image_slide import start_report_with_image_slide
 
 from flask import Blueprint
 
@@ -16,4 +16,4 @@ get_report(blueprint)
 import_report(blueprint)
 render_report(blueprint)
 start_report_with_pivot_table(blueprint)
-start_report_with_slide(blueprint)
+start_report_with_image_slide(blueprint)
